requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
faiss-cpu==1.7.4
filelock==3.13.1
filetype==1.2.0
Flask==3.0.1
fsspec==2023.12.2
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonify==0.5
keybert==0.8.3
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.3
openai-whisper==20231117
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
pydub==0.25.1
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
textblob==0.17.1
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.37.1
txtai==6.3.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
